Moved the products reloading and caching out of product service and made it a dependency of the service
